rename PlaceGuideWithUser to PlaceGuideWithUserPair
